+++ example/lib/main.dart
@@ -47,14 +47,9 @@ class _App extends StatelessWidget {
   const _App();
 
   void _test2() {
-    qq;
-    qw;
-    qe;
-    qr;
-    qqq("hello");
-    qqw("hello");
-    qqe("hello");
-    qqr("hello");
+    qqq("HF.debugShorterUS: ${HF.debugShorterUS}");
+    qqq("HF.debugShorterMS: ${HF.debugShorterMS}");
+    qqq("HF.debugShorterS: ${HF.debugShorterS}");
   }
 
   @override